Create SiteDetailsTable component
